refactor(paytoqueue): rename payTo queues filenames
