+++ src/connectors/queue/payTo/index.ts
@@ -1,2 +1,2 @@
-export * from './payToByMatters'
-export * from './payToByBlockchain'
+export * from './matters'
+export * from './blockchain'
